+++ src/components/common/tag.tsx
@@ -1,56 +1,45 @@
 'use client';
 
-type TagModalProps = {
-  isOpen: boolean;
-  onClose: () => void;
-  allTags: string[];
+type TagProps = {
   selectedTags: string[];
-  onTagChange: (tag: string) => void;
+  onSearchClick: () => void;
+  onResetClick: () => void;
 };
 
-const TagModal = ({
-  isOpen,
-  onClose,
-  allTags,
-  selectedTags,
-  onTagChange,
-}: TagModalProps) => {
-  if (!isOpen) return null;
+const Tag = ({ selectedTags, onSearchClick, onResetClick }: TagProps) => {
+  const hasSelectedTags = selectedTags.length > 0;
 
   return (
-    <div
-      className="fixed inset-0 bg-[#654321] flex flex-col items-center justify-center z-50 p-8"
-      style={{ backgroundColor: 'rgba(101, 67, 33, 1)' }} // #654321
-    >
-      <div className="w-full max-w-md text-[#F8F5E9]">
-        <h2 className="text-3xl font-bold mb-8 text-center">タグ検索</h2>
-        <div className="space-y-4 mb-10">
-          {allTags.map((tag) => (
-            <label key={tag} className="flex items-center text-2xl">
-              <input
-                type="checkbox"
-                checked={selectedTags.includes(tag)}
-                onChange={() => onTagChange(tag)}
-                className="appearance-none h-8 w-8 border-2 border-[#F8F5E9] rounded-sm bg-transparent checked:bg-[#F8F5E9] checked:border-transparent mr-4"
-              />
-              <span>{tag}</span>
-            </label>
-          ))}
-        </div>
-        <div className="flex justify-between items-center">
-          <button onClick={onClose} className="text-2xl">
-            戻る
-          </button>
+    <div className="text-center mb-8">
+      {hasSelectedTags ? (
+        <div>
           <button
-            onClick={onClose}
-            className="px-12 py-3 bg-[#F8F5E9] text-[#654321] text-2xl font-bold rounded-lg"
+            onClick={onResetClick}
+            className="inline-flex items-center justify-center h-[49px] px-[54px] py-2 text-white bg-accent rounded-sm text-body2 shadow-button text-center mb-4"
           >
-            検索
+            × タグをリセット
           </button>
+          <div className="flex justify-center gap-2 flex-wrap">
+            {selectedTags.map((tag) => (
+              <span
+                key={tag}
+                className="border border-accent text-accent px-3 py-1 rounded-full text-sm"
+              >
+                {tag}
+              </span>
+            ))}
+          </div>
         </div>
-      </div>
+      ) : (
+        <button
+          onClick={onSearchClick}
+          className="inline-flex items-center justify-center h-[49px] px-[54px] py-2 text-accent bg-base border-2 border-accent rounded-sm text-body2 shadow-button text-center"
+        >
+          タグ検索
+        </button>
+      )}
     </div>
   );
 };
 
-export default TagModal;
+export default Tag;
